+++ src/test/java/cn/howso/mybatis/mapper/UserMapper.java
@@ -4,5 +4,5 @@ import cn.howso.mybatis.anno.Table;
 import cn.howso.mybatis.model.Example;
 import cn.howso.mybatis.model.User;
 @Table(name="sys_user")
-public interface UserMapper extends BaseMapper<User, Example, Long>{
+public interface UserMapper extends BaseMapper<User, Example, Integer>{
 }
